moving icons from dataflow to shared-assets
